+++ src/cv_models/schemas.py
@@ -5,9 +5,9 @@ from enum import StrEnum
 
 class CVModelEnum(StrEnum):
 
+    YOLO8N = 'yolo8n',
     YOLO8S = 'yolo8s',
-    YOLO8M = 'yolo8m',
-    YOLO8N = 'yolo8n'
+    YOLO8M = 'yolo8m'
 
 class CVModelSchema(BaseModel):
